+++ src/nti/nikola_chameleon/viewlets.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+Holds viewlet manager interfaces defined by
+ZCML.
+
+There should be no actual contents in this file.
+
+.. versionadded:: 0.0.1a2
+
+"""
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function